refactor: rename a few test namespaces
